requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pgvector
pydantic
torch
sentence-transformers
redis
requests
python-dotenv
